+++ site/src/pages/preview/[domain]/[language]/admin/page.page.tsx
@@ -1,5 +1,5 @@
 import { BlockPreviewProvider, IFrameBridgeProvider, useIFrameBridge } from "@comet/cms-site";
-import { PageContentBlock } from "@src/documents/pages/blocks/PageContentBlock";
+import { PageContentBlock } from "@src/documents/blocks/PageContentBlock";
 import * as React from "react";
 
 const PreviewPage: React.FunctionComponent = () => {
